Version0_001/requirements-linux.txt
colorama==0.4.6
tqdm==4.66.4
pyyaml==6.0.1
numpy==1.26.4
scipy==1.11.4
sqlite-utils==3.36
sounddevice==0.4.6
espeakng==1.0.3
piper-tts==1.2.0
llama-cpp-python==0.2.90